Extract timer into module
